pyproject.toml
[project]
name = "local-ai-coordinator"
version = "0.1.0"
description = "阿里云端协调器 - 接收前端/APP请求，通过 NATS 转发给本地 AI Worker (local-ai-worker)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "郭浩" }
]
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "nats-py>=2.14.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.45.0",
]
